+++ AutomatedLabCore/functions/Internals/Remove-LabDeploymentFiles.ps1
@@ -1,16 +1,8 @@
 ﻿function Remove-LabDeploymentFiles
 {
 
-    Invoke-LabCommand -ComputerName (Get-LabVM) -ActivityName 'Remove deployment files (files used during deployment)' -AsJob -NoDisplay -ScriptBlock `
-    {
-        $paths = 'C:\Unattend.xml',
-            'C:\WSManRegKey.reg',
-            'C:\AdditionalDisksOnline.ps1',
-            'C:\WinRmCustomization.ps1',
-            'C:\DeployDebug',
-            'C:\ALLibraries',
-            "C:\$($env:COMPUTERNAME).cer"
-            
+    Invoke-LabCommand -ComputerName (Get-LabVM) -ActivityName 'Remove deployment files (files used during deployment)' -AsJob -NoDisplay -ScriptBlock {
+        $paths = 'C:\Unattend.xml', $ExecutionContext.InvokeCommand.ExpandString($AL_DeployDebugFolder)
         $paths | Remove-Item -Force -Recurse -ErrorAction SilentlyContinue
-    }
+    } -Variable (Get-Variable -Scope Global -Name AL_DeployDebugFolder)
 }